requirements.txt
beautifulsoup4
requests
boto3
python-dotenv